+++ __snapshots__/github.js
@@ -857,3 +857,373 @@ exports['GitHub pullRequestIterator finds merged pull requests with labels 1'] =
     ]
   }
 ]
+
+exports['GitHub pullRequestIterator handles merged pull requests without files 1'] = [
+  {
+    "sha": "198808aae0adc62c9ad95452170fd691ac40bd00",
+    "number": 2827,
+    "baseBranchName": "main",
+    "headBranchName": "fix2796",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "fix: owl bot endlessly retried failed owl-bot-lock builds",
+    "body": "Fixes https://github.com/googleapis/repo-automation-bots/issues/2796\r\n",
+    "files": []
+  },
+  {
+    "sha": "2838dadb5214f5dd3cfa85665428aa4b923a29e8",
+    "number": 2824,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-trusted-contribution",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update trusted-contribution deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/trusted-contribution",
+    "files": [
+      "packages/trusted-contribution/package-lock.json",
+      "packages/trusted-contribution/package.json"
+    ]
+  },
+  {
+    "sha": "b689b19c4aa839173d5e1d09fe8e3d271a18ff45",
+    "number": 2822,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-sync-repo-settings",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update sync-repo-settings deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/sync-repo-settings",
+    "files": [
+      "packages/sync-repo-settings/package-lock.json",
+      "packages/sync-repo-settings/package.json"
+    ]
+  },
+  {
+    "sha": "d270afdf45d6db550003c00725aea0294d439897",
+    "number": 2821,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-secret-rotator",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update secret-rotator deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/secret-rotator",
+    "files": [
+      "packages/secret-rotator/package-lock.json",
+      "packages/secret-rotator/package.json"
+    ]
+  },
+  {
+    "sha": "0b5d2853f75230528c2db9322a8a035219fc3e06",
+    "number": 2820,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-release-please",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update release-please deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/release-please",
+    "files": [
+      "packages/release-please/package-lock.json",
+      "packages/release-please/package.json"
+    ]
+  },
+  {
+    "sha": "9faf33d20beb160b257e368a69c425989fae4d1d",
+    "number": 2819,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-policy",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update policy deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/policy",
+    "files": [
+      "packages/policy/package-lock.json"
+    ]
+  },
+  {
+    "sha": "2c8c92d3f154365ca968afa9ac56f0cac3211e18",
+    "number": 2818,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-release-brancher",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update release-brancher deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/release-brancher",
+    "files": [
+      "packages/release-brancher/package-lock.json",
+      "packages/release-brancher/package.json"
+    ]
+  },
+  {
+    "sha": "58f88085907b0d54efd9f8499bebddbaba08432f",
+    "number": 2817,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-object-selector",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update object-selector deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/object-selector",
+    "files": [
+      "packages/object-selector/package-lock.json",
+      "packages/object-selector/package.json"
+    ]
+  },
+  {
+    "sha": "9390299a1cbb157939309e6682abb7785df2c7a2",
+    "number": 2815,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-mono-repo-publish",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update mono-repo-publish deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/mono-repo-publish",
+    "files": [
+      "packages/mono-repo-publish/package-lock.json"
+    ]
+  },
+  {
+    "sha": "11f32f8b4613618d3ea9aa1cbd482c3b8e47b5ce",
+    "number": 2814,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-loadtest-bot",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update loadtest-bot deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/loadtest-bot",
+    "files": [
+      "packages/loadtest-bot/package-lock.json",
+      "packages/loadtest-bot/package.json"
+    ]
+  },
+  {
+    "sha": "c09a32eccb170f772e3dc69c6a035f9fc350a6ef",
+    "number": 2813,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-label-utils",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update label-utils deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/label-utils",
+    "files": [
+      "packages/label-utils/package-lock.json",
+      "packages/label-utils/package.json"
+    ]
+  },
+  {
+    "sha": "5a30de169af5376422151806c979d2fc2b3ca3d1",
+    "number": 2812,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-label-sync",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update label-sync deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/label-sync",
+    "files": [
+      "packages/label-sync/package-lock.json",
+      "packages/label-sync/package.json"
+    ]
+  },
+  {
+    "sha": "34e061388fefc4e5143319233d9fbbc3f9b11a87",
+    "number": 2811,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-generate-bot",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update generate-bot deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/generate-bot",
+    "files": [
+      "packages/generate-bot/package-lock.json"
+    ]
+  },
+  {
+    "sha": "6ef6b701d975f7ad32be5747f806969eb51d9cf2",
+    "number": 2810,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-header-checker-lint",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update header-checker-lint deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/header-checker-lint",
+    "files": [
+      "packages/header-checker-lint/package-lock.json",
+      "packages/header-checker-lint/package.json"
+    ]
+  },
+  {
+    "sha": "350a9c3109d845124b7f43d9d3f96dd1903cd3fa",
+    "number": 2791,
+    "baseBranchName": "main",
+    "headBranchName": "fix-owlbot",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "fix(owl-bot): limit organizations for syncLabels",
+    "body": "It mitigates #2750\r\n",
+    "files": [
+      "packages/owl-bot/src/owl-bot.ts",
+      "packages/owl-bot/test/owl-bot.ts"
+    ]
+  },
+  {
+    "sha": "7dc3495327012e5b2ab3c93df6abaaf35e08299d",
+    "number": 2788,
+    "baseBranchName": "main",
+    "headBranchName": "renovate/node-16.x",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "chore(deps): update dependency @types/node to v16",
+    "body": "[![WhiteSource Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)\n\nThis PR contains the following updates:\n\n| Package | Change | Age | Adoption | Passing | Confidence |\n|---|---|---|---|---|---|\n| [@types/node](https://togithub.com/DefinitelyTyped/DefinitelyTyped) | [`^14.0.0` -> `^16.0.0`](https://renovatebot.com/diffs/npm/@types%2fnode/14.17.4/16.11.6) | [![age](https://badges.renovateapi.com/packages/npm/@types%2fnode/16.11.6/age-slim)](https://docs.renovatebot.com/merge-confidence/) | [![adoption](https://badges.renovateapi.com/packages/npm/@types%2fnode/16.11.6/adoption-slim)](https://docs.renovatebot.com/merge-confidence/) | [![passing](https://badges.renovateapi.com/packages/npm/@types%2fnode/16.11.6/compatibility-slim/14.17.4)](https://docs.renovatebot.com/merge-confidence/) | [![confidence](https://badges.renovateapi.com/packages/npm/@types%2fnode/16.11.6/confidence-slim/14.17.4)](https://docs.renovatebot.com/merge-confidence/) |\n\n---\n\n### Configuration\n\n📅 **Schedule**: \"after 9am and before 3pm\" (UTC).\n\n🚦 **Automerge**: Disabled by config. Please merge this manually once you are satisfied.\n\n♻ **Rebasing**: Renovate will not automatically rebase this PR, because other commits have been found.\n\n🔕 **Ignore**: Close this PR and you won't be reminded about this update again.\n\n---\n\n - [ ] <!-- rebase-check -->If you want to rebase/retry this PR, click this checkbox.\n\n---\n\nThis PR has been generated by [WhiteSource Renovate](https://renovate.whitesourcesoftware.com). View repository job log [here](https://app.renovatebot.com/dashboard#github/googleapis/repo-automation-bots).",
+    "files": [
+      "packages/release-trigger/package-lock.json",
+      "packages/release-trigger/package.json"
+    ]
+  },
+  {
+    "sha": "d2f43552140f5791f4bcd307a91cd1cf9a0c4e52",
+    "number": 2785,
+    "baseBranchName": "main",
+    "headBranchName": "fix-failing-tests",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "fix: tests failed in github actions",
+    "body": "Fixes https://github.com/googleapis/repo-automation-bots/issues/2783",
+    "files": [
+      "packages/owl-bot/test/commit-post-processor-update.ts"
+    ]
+  },
+  {
+    "sha": "4a7da3aec226612ace78b14e097875ceee67f2bb",
+    "number": 2781,
+    "baseBranchName": "main",
+    "headBranchName": "only-main",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "fix(owlbot): only run against main branch unless labeled",
+    "body": "Avoid running OwlBot unless `owlbot:run` is added, if the PR is not against the default branch.\r\n\r\nFixes #2745",
+    "files": [
+      "packages/owl-bot/src/owl-bot.ts",
+      "packages/owl-bot/test/owl-bot.ts"
+    ]
+  },
+  {
+    "sha": "bf01ffc4f45bc8f9de21d6d9c0b7c6bccf72fe70",
+    "number": 2780,
+    "baseBranchName": "main",
+    "headBranchName": "after-token",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "fix: install credentials *after* running the container",
+    "body": "",
+    "files": [
+      "packages/owl-bot/cloud-build/update-pr.yaml",
+      "packages/owl-bot/src/bin/commands/commit-post-processor-update.ts"
+    ]
+  },
+  {
+    "sha": "64184679e1630d33eb6d55f45e442a7cb518e976",
+    "number": 2779,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-sync-repo-settings",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update sync-repo-settings deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/sync-repo-settings",
+    "files": [
+      "packages/sync-repo-settings/package-lock.json",
+      "packages/sync-repo-settings/package.json"
+    ]
+  },
+  {
+    "sha": "d1edf2bac6e7df061b9e70fedbc81e2fcd1bdde1",
+    "number": 2778,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-trusted-contribution",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update trusted-contribution deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/trusted-contribution",
+    "files": [
+      "packages/trusted-contribution/package-lock.json",
+      "packages/trusted-contribution/package.json"
+    ]
+  },
+  {
+    "sha": "9205148b7cd700ee93669307953d4d737965126b",
+    "number": 2777,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-snippet-bot",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update snippet-bot deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/snippet-bot",
+    "files": [
+      "packages/snippet-bot/package-lock.json"
+    ]
+  },
+  {
+    "sha": "91d63b013a30d6ca90e97f17d475d32dfc7a5838",
+    "number": 2776,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-secret-rotator",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update secret-rotator deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/secret-rotator",
+    "files": [
+      "packages/secret-rotator/package-lock.json",
+      "packages/secret-rotator/package.json"
+    ]
+  },
+  {
+    "sha": "d37ef994fb1084fb93cb0f2b359da4f4b8c1bd01",
+    "number": 2775,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-release-please",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update release-please deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/release-please",
+    "files": [
+      "packages/release-please/package-lock.json",
+      "packages/release-please/package.json"
+    ]
+  },
+  {
+    "sha": "785c5a6248c596d21f43d596450413021690fb0f",
+    "number": 2774,
+    "baseBranchName": "main",
+    "headBranchName": "update-deps-release-brancher",
+    "labels": [
+      "cla: yes"
+    ],
+    "title": "deps: update release-brancher deps",
+    "body": "ran ./scripts/update-dependencies.sh packages/release-brancher",
+    "files": [
+      "packages/release-brancher/package-lock.json",
+      "packages/release-brancher/package.json"
+    ]
+  }
+]